moved detailedCompare to adapter utils
