+++ packages/adapter-utils/index.ts
@@ -16,3 +16,4 @@
 export * from './src/utils'
 export * from './src/change_validator'
 export * from './src/deploy'
+export * from './src/compare'
